# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "نظام المكتبة الجامعية",
        "description": "توثيق API لنظام المكتبة الجامعية",
        "contact": {
            "email": "admin@university.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "تسجيل الدخول للطالب",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "student_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": " يوسف طارق احمد"
                                    },
                                    "student_id": {
                                        "type": "numeric",
                                        "example": 15901
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تسجيل الدخول بنجاح."
                    },
                    "401": {
                        "description": "بيانات الطالب غير صحيحة."
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "تسجيل خروج الطالب",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "تم تسجيل الخروج بنجاح."
                    },
                    "401": {
                        "description": "غير مصرح لك."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "الحصول على بيانات الطالب الحالي",
                "description": "إرجاع بيانات الطالب المسجل دخوله حالياً",
                "operationId": "8e0149b7f62d4e6ff6573491d62b4560",
                "responses": {
                    "200": {
                        "description": "بيانات الطالب",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "يوسف طارق احمد"
                                        },
                                        "student_id": {
                                            "type": "integer",
                                            "example": 15901
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "http://localhost/storage/students/image.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح لك"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "عرض قائمة التصنيفات",
                "description": "عرض قائمة بجميع التصنيفات المتاحة",
                "operationId": "1ba94c3ddf4dbb87208e46c12a5d1a45",
                "responses": {
                    "200": {
                        "description": "قائمة التصنيفات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "الإدارة"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories/{name}/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "عرض الكتب حسب التصنيف",
                "description": "عرض قائمة الكتب في التصنيف المحدد",
                "operationId": "718431eacec5659ee2067eebc435c9b7",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "اسم التصنيف (مثال: الإدارة)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "قائمة الكتب في التصنيف المحدد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Book"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح"
                    },
                    "404": {
                        "description": "التصنيف غير موجود"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/search/history": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "عرض سجل البحث",
                "description": "عرض آخر 10 عمليات بحث قام بها الطالب",
                "operationId": "60247dd649826ae9aa186e92028a3bbb",
                "responses": {
                    "200": {
                        "description": "سجل البحث",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "search_history": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "search_term": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/search/history/{id}": {
            "delete": {
                "tags": [
                    "Search"
                ],
                "summary": "حذف عملية بحث من السجل",
                "description": "حذف عملية بحث محددة من سجل البحث",
                "operationId": "68add3ad16627e0c5e7606a0a4ca8ad3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "معرف عملية البحث",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم الحذف بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "تم حذف سجل البحث بنجاح"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح"
                    },
                    "404": {
                        "description": "لم يتم العثور على سجل البحث"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "عنوان الكتاب"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 5
                    },
                    "category": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "authors": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}